+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfFuror.java
@@ -49,7 +49,7 @@ public class RingOfFuror extends Ring {
 		}
 	}
 
-	public String upgradeStat1(int level){
+	public String upgradeStat1(long level){
 		if (cursed && cursedKnown) level = Math.min(-1, level-3);
 		return Messages.decimalFormat("#.##", 100f * (Math.pow(1.09051f, level+1)-1f)) + "%";
 	}
