+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfFrost.java
@@ -113,7 +113,7 @@ public class WandOfFrost extends DamageWand {
 	}
 
 	@Override
-	public String upgradeStat2(int level) {
+	public String upgradeStat2(long level) {
 		return Integer.toString(2 + level);
 	}
