+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Dagger.java
@@ -106,7 +106,7 @@ public class Dagger extends MeleeWeapon {
 	}
 
 	@Override
-	public String upgradeAbilityStat(int level) {
+	public String upgradeAbilityStat(long level) {
 		return Integer.toString(2+level);
 	}
