+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Gloves.java
@@ -109,7 +109,7 @@ public class Gloves extends MeleeWeapon {
 		}
 	}
 
-	public String upgradeAbilityStat(int level){
+	public String upgradeAbilityStat(long level){
 		return "+" + augment.damageFactor(2 + Math.round(0.5f*level));
 	}
