+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Sword.java
@@ -87,7 +87,7 @@ public class Sword extends MeleeWeapon {
 		}
 	}
 
-	public String upgradeAbilityStat(int level){
+	public String upgradeAbilityStat(long level){
 		int dmgBoost = 5 + level;
 		return augment.damageFactor(min(level)+dmgBoost) + "-" + augment.damageFactor(max(level)+dmgBoost);
 	}
